# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PrintFlow API - Moran Creative",
        "description": "API para gestión de pedidos de impresión 3D",
        "contact": {
            "email": "admin@morancreative.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/materials": {
            "get": {
                "tags": [
                    "Materiales"
                ],
                "summary": "Listar materiales con colores (público)",
                "operationId": "ed04a96cf463463a497cf4d45b931cd1",
                "responses": {
                    "200": {
                        "description": "Lista de materiales activos con sus colores disponibles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Materiales - Admin"
                ],
                "summary": "Crear material (solo admin)",
                "operationId": "9c9de3177d10f03d75f9573371f9186f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "density_g_cm3",
                                    "price_per_gram"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PLA"
                                    },
                                    "density_g_cm3": {
                                        "type": "number",
                                        "example": 1.24
                                    },
                                    "price_per_gram": {
                                        "type": "number",
                                        "example": 2
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Material creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/materials/{id}": {
            "put": {
                "tags": [
                    "Materiales - Admin"
                ],
                "summary": "Actualizar material (solo admin)",
                "operationId": "2bc3ff8daffbd0e30f6673202d857344",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PLA+"
                                    },
                                    "density_g_cm3": {
                                        "type": "number",
                                        "example": 1.24
                                    },
                                    "price_per_gram": {
                                        "type": "number",
                                        "example": 2.5
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Material actualizado"
                    },
                    "404": {
                        "description": "Material no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Materiales - Admin"
                ],
                "summary": "Eliminar material (solo admin)",
                "operationId": "9c353fdc77702c2abb6f63783f4e44f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material eliminado"
                    },
                    "404": {
                        "description": "Material no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/materials/{id}/colors": {
            "post": {
                "tags": [
                    "Materiales - Admin"
                ],
                "summary": "Agregar color a un material (solo admin)",
                "operationId": "a4a0008d77967e2fce0fd11acca326dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "color_name",
                                    "hex_code"
                                ],
                                "properties": {
                                    "color_name": {
                                        "type": "string",
                                        "example": "Rojo"
                                    },
                                    "hex_code": {
                                        "type": "string",
                                        "example": "#DC2626"
                                    },
                                    "extra_cost": {
                                        "type": "number",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Color agregado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/materials/{id}/colors/{colorId}": {
            "delete": {
                "tags": [
                    "Materiales - Admin"
                ],
                "summary": "Eliminar color de un material (solo admin)",
                "operationId": "b7d5bbb3ec1ffacd99562c5e91055119",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "colorId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Color eliminado"
                    },
                    "404": {
                        "description": "Color no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}